requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
clamd==1.0.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2
django-clamd==0.4.0
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==22.1
django-hashid-field==3.3.7
django-localflavor==3.1
django-model-utils==4.3.1
django-oauth-toolkit==2.2.0
django-templated-mail==1.1.1
django-vies==6.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-social-oauth2==1.2.1
ecdsa==0.18.0
hashids==1.3.1
idna==3.4
importlib-metadata==6.0.0
isodate==0.6.1
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.4.2
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.1
oauthlib==3.2.2
Pillow==9.3.0
placebo==0.9.0
platformdirs==2.6.2
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.6.0
python-dotenv==0.21.0
python-jose==3.3.0
python-stdnum==1.18
python3-openid==3.2.0
pytz==2022.7
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
retrying==1.3.4
rsa==4.9
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
wcwidth==0.2.5
whitenoise==6.2.0
wrapt==1.14.1
zeep==4.2.1
zipp==3.11.0
